+++ src/basilisp/sitecustomize.py
@@ -1,0 +1,3 @@
+import basilisp.main
+
+basilisp.main.init()