+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  * Added support for passing through `:tag` metadata to the generated Python AST (#354)
  * Added support for calling symbols as functions on maps and sets (#775)
  * Added support for passing command line arguments to Basilisp (#779)
- * Added support for autocompleting names in the `python/` pseudo-namespace for Python builtins at the REPL (#787) 
+ * Added support for autocompleting names in the `python/` pseudo-namespace for Python builtins at the REPL (#787)
+ * Added a subcommand for bootstrapping the Python installation with Basilisp (#7??)
 
 ### Changed
  * Optimize calls to Python's `operator` module into their corresponding native operators (#754)